Fix membership dedup ordering and strip comments from kafka DSM membership reporting
Only mark a membership as reported once reportConsumerGroupMember
actually attempts the report (i.e. cluster id is known), instead of
marking it unconditionally before the cluster-id check. Also aligns
generationId serialization with writeInt/unpackInt used elsewhere for
int fields, and removes redundant/duplicated comments introduced by
the feature.

+++ dd-trace-core/src/main/java/datadog/trace/core/datastreams/MsgPackDatastreamsPayloadWriter.java
@@ -293,7 +293,6 @@ public class MsgPackDatastreamsPayloadWriter implements DatastreamsPayloadWriter
     packer.writeUTF8(CONFIGS);
     packer.startArray(configs.size());
     for (KafkaConfigReport config : configs) {
-      // Type, KafkaClusterId, ConsumerGroup, MemberId, GenerationId, MemberProtocol, Config
       packer.startMap(7);
 
       packer.writeUTF8(CONFIG_TYPE);
@@ -309,7 +308,7 @@ public class MsgPackDatastreamsPayloadWriter implements DatastreamsPayloadWriter
       packer.writeString(config.getMemberId(), null);
 
       packer.writeUTF8(CONFIG_GENERATION_ID);
-      packer.writeLong(config.getGenerationId());
+      packer.writeInt(config.getGenerationId());
 
       packer.writeUTF8(CONFIG_MEMBER_PROTOCOL);
       packer.writeString(config.getMemberProtocol(), null);
